% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/scale_sf.R
\name{qtot}
\alias{qtot}
\title{Sap flux scaling}
\usage{
qtot(v, a = 0, b, woodType = c("Tracheid", "Diffuse-porous", "Ring-porous"),
  uncertainty = FALSE, nboot = 5000, treeRadius, sapRadius = NULL)
}
\arguments{
\item{v}{sap flux density observations \eqn{(g / m^2 / time)}}

\item{a}{distance from cambium (in \eqn{m}) for outer end of sap flux probe (default = 0)}

\item{b}{distance from cambium (in \eqn{m}) for inner end of sap flux probe}

\item{woodType}{xylem type for radial profile selection}

\item{uncertainty}{estimate with uncertainty (\code{TRUE}) or use mean parameters (\code{FALSE}, default)}

\item{nboot}{number of draws from posterior distribution to use for prediction}

\item{treeRadius}{tree radius minus bark (in \eqn{m})}

\item{sapRadius}{sapwood depth (in \eqn{m}) (optional)}
}
\value{
The integral of sap flux density throughout a tree trunk. Output is a matrix of predictions (number of observations x number of bootstrap samples). If "uncertainty" is \code{FALSE}, returns a vector of mean values.
}
\description{
This function generates predictions of whole tree sap flow based on sap flux density observations.
}
\examples{
qtot(egtree, 0, 0.02, "Ring-porous", FALSE, 1, 0.0255, NULL)
}

