\name{sapflux-package}
\alias{sapflux-package}
\alias{sapflux}
\docType{package}
\title{
\packageTitle{sapflux}
}
\description{
\packageDescription{sapflux}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sapflux}
\packageIndices{sapflux}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sapflux}

Maintainer: \packageMaintainer{sapflux}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
