% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sapflux.R
\docType{package}
\name{sapflux}
\alias{sapflux}
\alias{sapflux-package}
\title{sapflux: A package for calculating whole tree water use.}
\description{
The package currently contains one main function:
qtot
}
\section{qtot}{

The qtot function generates estimates of whole tree water use with measurements of sap flux density from thermal dissipation probes.get
}

